rate limit deletion of UUIDs

+++ src/services/admin/user-verifications.js
@@ -1,4 +1,8 @@
-import { UserVerifications } from "../../init.js";
+import {
+  UserVerifications,
+  DailyVerificationCount,
+  DailyVerificationDeletions,
+} from "../../init.js";
 import { logWithTimestamp } from "../../utils/utils.js";
 
 async function getUserVerification(req, res) {
@@ -49,11 +53,40 @@ async function deleteUserVerification(req, res) {
       return res.status(400).json({ error: "No UUID provided." });
     }
 
+    // Limit the number of deletions per day to 2% of the number of verifications per day
+    const verificationCountDoc = await DailyVerificationCount.findOne({
+      date: new Date().toISOString().slice(0, 10),
+    }).exec();
+    // NOTE: If we add other verification providers, we need to update the following line
+    const sessionCountToday =
+      verificationCountDoc.veriff.sessionCount +
+      (verificationCountDoc?.vouched?.jobCount ?? 0);
+    const deletionCountDoc = await DailyVerificationDeletions.findOne({
+      date: new Date().toISOString().slice(0, 10),
+    }).exec();
+    const deletionCountToday = deletionCountDoc.deletionCount;
+    if (deletionCountToday >= sessionCountToday * 0.02) {
+      logWithTimestamp(
+        "DELETE admin/user-verifications: Deletion limit reached for today. Exiting."
+      );
+      return res
+        .status(429)
+        .json({ error: "Deletion limit reached for today. Try again tomorrow." });
+    }
+
     const result = await UserVerifications.deleteOne({ "govId.uuid": uuid }).exec();
     if (result.acknowledged && result.deletedCount >= 1) {
       logWithTimestamp(
         `DELETE admin/user-verifications: Deleted user with UUID ${uuid}.`
       );
+
+      // Increment the deletion count for today
+      await DailyVerificationDeletions.updateOne(
+        { date: new Date().toISOString().slice(0, 10) },
+        { $inc: { deletionCount: 1 } },
+        { upsert: true }
+      ).exec();
+
       return res.status(200).json({ message: "User deleted" });
     } else {
       logWithTimestamp(
